+++ src/main/java/com/raoulvdberge/refinedstorage/render/BakedModelPortableGrid.java
@@ -27,7 +27,7 @@ public class BakedModelPortableGrid implements IPerspectiveAwareModel {
     public Pair<? extends IBakedModel, Matrix4f> handlePerspective(ItemCameraTransforms.TransformType cameraTransformType) {
         TRSRTransformation transform = RSUtils.getDefaultBlockTransforms().get(cameraTransformType);
 
-        return Pair.of(this, transform == null ? RSUtils.EMPTY_MATRIX : transform.getMatrix());
+        return Pair.of(this, transform == null ? RSUtils.EMPTY_MATRIX_TRANSFORM : transform.getMatrix());
     }
 
     @Override